pyproject.toml
[project]
name = "langchain1-0"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "chromadb>=1.3.5",
    "deepagents>=0.2.8",
    "langchain>=1.1.0",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-deepseek>=1.0.1",
    "langchain-huggingface>=1.1.0",
    "langchain-ollama>=0.1.0",
    "langchain-tavily>=0.2.13",
    "langgraph-cli[inmem]>=0.4.7",
    "pypdf>=6.4.0",
    "sentence-transformers>=5.1.2",
    "tavily-python>=0.7.13",
    "torch>=2.9.1",
    "torchaudio>=2.9.1",
    "torchvision>=0.24.1",
    "unstructured>=0.18.21",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
